// client/src/state/actionTypes/noteActionTypes.ts
export enum NoteActionTypes {
  NOTE_REQUEST = "NOTE_REQUEST",
  NOTE_SUCCESS = "NOTE_SUCCESS",
  NOTE_FAILURE = "NOTE_FAILURE",
  NOTE_GET_ONE_REQUEST = "NOTE_GET_ONE_REQUEST",
  NOTE_GET_ONE_SUCCESS = "NOTE_GET_ONE_SUCCESS",
  NOTE_GET_ONE_FAILURE = "NOTE_GET_ONE_FAILURE",
  NOTE_GET_ONE_RESET = "NOTE_GET_ONE_RESET",
  NOTE_ADD_REQUEST = "NOTE_ADD_REQUEST",
  NOTE_ADD_SUCCESS = "NOTE_ADD_SUCCESS",
  NOTE_ADD_FAILURE = "NOTE_ADD_FAILURE",
  NOTE_ADD_RESET = "NOTE_ADD_RESET",
  NOTE_EDIT_REQUEST = "NOTE_EDIT_REQUEST",
  NOTE_EDIT_SUCCESS = "NOTE_EDIT_SUCCESS",
  NOTE_EDIT_FAILURE = "NOTE_EDIT_FAILURE",
  NOTE_EDIT_RESET = "NOTE_EDIT_RESET",
  NOTE_DELETE_REQUEST = "NOTE_DELETE_REQUEST",
  NOTE_DELETE_SUCCESS = "NOTE_DELETE_SUCCESS",
  NOTE_DELETE_FAILURE = "NOTE_DELETE_FAILURE",
  NOTE_DELETE_RESET = "NOTE_DELETE_RESET",
}